+++ src/xrEngine/line_editor.h
@@ -16,7 +16,7 @@ namespace text_editor
 class line_editor : public IInputReceiver
 {
 public:
-    line_editor(u32 str_buffer_size);
+    line_editor(size_t str_buffer_size);
     virtual ~line_editor() = default;
 
     IC line_edit_control& control() { return m_control; }
